@@ -1,12 +1,12 @@
 {
   "name": "velocity-react",
-  "version": "1.3.3",
+  "version": "1.4.0",
   "description": "React components to wrap Velocity animations",
   "main": "index.js",
   "scripts": {
     "compile": "babel src --out-dir .",
     "demo": "webpack-dev-server --progress --colors --content-base demo --hot --inline --config demo/webpack.config.js",
-    "prepublish": "npm run compile"
+    "prepare": "npm run compile"
   },
   "repository": {
     "type": "git",
@@ -27,9 +27,9 @@
     "*.js": "eslint --quiet"
   },
   "dependencies": {
-    "lodash": "^3.10.1",
+    "lodash": "^4.17.5",
     "prop-types": "^15.5.8",
-    "react-transition-group": "^1.1.2",
+    "react-transition-group": "^2.0.0",
     "velocity-animate": "^1.4.0"
   },
   "peerDependencies": {
@@ -55,10 +55,8 @@
     "husky": "^0.14.3",
     "lint-staged": "^4.2.3",
     "prettier": "^1.7.0",
-    "react": "^15.5.0",
-    "react-dom": "^15.5.0",
-    "react-hot-loader": "^1.2.9",
-    "react-tween-state": "^0.1.3",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "style-loader": "^0.12.3",
     "twemoji": "^1.4.1",
     "underscore.string": "^3.2.0",
