@@ -1,7 +1,6 @@
 var React = require('react');
 var createReactClass = require('create-react-class');
 var VelocityComponent = require('../../src/velocity-component');
-var tweenState = require('react-tween-state');
 var s = require('underscore.string');
 
 var Box = require('../components/box');
@@ -38,9 +37,7 @@ var EFFECTS = [
 var ToggleBox = createReactClass({
   displayName: 'ToggleBox',
 
-  mixins: [ tweenState.Mixin ],
-
-  getInitialState: function () {
+  getInitialState: function() {
     return {
       effect: EFFECTS[0],
       isIn: true,
@@ -49,87 +46,138 @@ var ToggleBox = createReactClass({
     };
   },
 
-  componentWillMount: function () {
+  componentDidMount: function() {
     // Tweening shows how animations work given re-renders
     this.startTweening();
   },
 
-  startTweening: function () {
-    this.tweenState('counter', {
-      duration: 5000,
-      endValue: 100,
-      onEnd: this.reverseTweening,
-    });
+  componentWillUnmount: function() {
+    this.stopTweening();
   },
 
-  reverseTweening: function () {
-    this.tweenState('counter', {
-      duration: 5000,
-      endValue: 0,
-      onEnd: this.startTweening,
-    });
+  startTweening: function() {
+    this.tweenInterval = setInterval(() => {
+      if (this.state.counter === 99) {
+        this.setState({ counter: 0 });
+      } else {
+        this.setState({ counter: this.state.counter + 1 });
+      }
+    }, 100);
   },
 
-  whenToggleClicked: function () {
+  stopTweening: function() {
+    clearInterval(this.tweenInterval);
+  },
+
+  whenToggleClicked: function() {
     this.setState({
       isIn: !this.state.isIn,
     });
   },
 
-  whenSelectChanged: function (evt) {
+  whenSelectChanged: function(evt) {
     this.setState({
       effect: evt.target.value,
       isIn: true,
     });
   },
 
-  whenOptionChanged: function (evt) {
+  whenOptionChanged: function(evt) {
     this.setState({ interruptBehavior: evt.target.value });
   },
 
-  render: function () {
-    var animation = 'transition.' + this.state.effect + (this.state.isIn ? 'In' : 'Out');
+  render: function() {
+    var animation =
+      'transition.' + this.state.effect + (this.state.isIn ? 'In' : 'Out');
 
     return (
       <div className="flex-box flex-column flex-1 align-items-center">
         <div>
-          <select value={this.state.effect} onChange={this.whenSelectChanged}>{this.renderEffects()}</select>
+          <select value={this.state.effect} onChange={this.whenSelectChanged}>
+            {this.renderEffects()}
+          </select>
         </div>
         <div className="flex-1">
-          <Box className="flex-1 flex-box flex-column align-items-center" style={{backgroundColor: '#f5f5f5'}} onClick={this.whenToggleClicked} instruction="Click!">
+          <Box
+            className="flex-1 flex-box flex-column align-items-center"
+            style={{ backgroundColor: '#f5f5f5' }}
+            onClick={this.whenToggleClicked}
+            instruction="Click!"
+          >
             {/*
               Use of key here keeps the component (and its set styles) from persisting across effects.
               Avoids flashing when switching effects.
             */}
-            <VelocityComponent key={this.state.effect} animation={animation} interruptBehavior={this.state.interruptBehavior}>
-              <Box>{Math.floor(this.getTweeningValue('counter'))}</Box>
+            <VelocityComponent
+              key={this.state.effect}
+              animation={animation}
+              interruptBehavior={this.state.interruptBehavior}
+            >
+              <Box>{this.state.counter}</Box>
             </VelocityComponent>
           </Box>
         </div>
-        <div style={{ fontStyle: 'italic', fontSize: '11px', padding: '0px 30px 10px', textAlign: 'center' }}>
-          Number counting to show the [non-]effects of rapid re-rendering on the animation.
+        <div
+          style={{
+            fontStyle: 'italic',
+            fontSize: '11px',
+            padding: '0px 30px 10px',
+            textAlign: 'center',
+          }}
+        >
+          Number counting to show the [non-]effects of rapid re-rendering on the
+          animation.
         </div>
         <div style={{ fontSize: 12 }}>
-          <h4 style={{ marginBottom: 4, fontWeight: 'bold', textAlign: 'center'}}>Interruption Behavior</h4>
+          <h4
+            style={{ marginBottom: 4, fontWeight: 'bold', textAlign: 'center' }}
+          >
+            Interruption Behavior
+          </h4>
           <label>
-            <input type="radio" name="interruptBehavior" value="stop" checked={this.state.interruptBehavior === 'stop'} onChange={this.whenOptionChanged}/> Stop
+            <input
+              type="radio"
+              name="interruptBehavior"
+              value="stop"
+              checked={this.state.interruptBehavior === 'stop'}
+              onChange={this.whenOptionChanged}
+            />{' '}
+            Stop
           </label>
           &nbsp;&nbsp;
           <label>
-            <input type="radio" name="interruptBehavior" value="finish" checked={this.state.interruptBehavior === 'finish'} onChange={this.whenOptionChanged}/> Finish
+            <input
+              type="radio"
+              name="interruptBehavior"
+              value="finish"
+              checked={this.state.interruptBehavior === 'finish'}
+              onChange={this.whenOptionChanged}
+            />{' '}
+            Finish
           </label>
           &nbsp;&nbsp;
           <label>
-            <input type="radio" name="interruptBehavior" value="queue" checked={this.state.interruptBehavior === 'queue'} onChange={this.whenOptionChanged}/> Queue
+            <input
+              type="radio"
+              name="interruptBehavior"
+              value="queue"
+              checked={this.state.interruptBehavior === 'queue'}
+              onChange={this.whenOptionChanged}
+            />{' '}
+            Queue
           </label>
         </div>
       </div>
     );
   },
 
-  renderEffects: function () {
-    return EFFECTS.map(function (effect) {
-      return (<option key={effect} value={effect}>{s.titleize(s.humanize(effect))}</option>);
+  renderEffects: function() {
+    return EFFECTS.map(function(effect) {
+      return (
+        <option key={effect} value={effect}>
+          {s.titleize(s.humanize(effect))}
+        </option>
+      );
     });
   },
 });
