@@ -39,10 +39,10 @@ Methods:
 /* eslint react/no-find-dom-node: 0 */
 
 var _ = {
-  forEach: require('lodash/collection/forEach'),
-  isEqual: require('lodash/lang/isEqual'),
-  keys: require('lodash/object/keys'),
-  omit: require('lodash/object/omit'),
+  forEach: require('lodash/forEach'),
+  isEqual: require('lodash/isEqual'),
+  keys: require('lodash/keys'),
+  omit: require('lodash/omit'),
 };
 
 var React = require('react');
@@ -71,11 +71,11 @@ class VelocityComponent extends React.Component {
     }
   }
 
-  componentWillUpdate(newProps) {
-    if (!_.isEqual(newProps.animation, this.props.animation)) {
-      if (newProps.interruptBehavior === 'stop') {
+  componentDidUpdate(oldProps) {
+    if (!_.isEqual(oldProps.animation, this.props.animation)) {
+      if (this.props.interruptBehavior === 'stop') {
         this._stopAnimation();
-      } else if (newProps.interruptBehavior === 'finish') {
+      } else if (this.props.interruptBehavior === 'finish') {
         this._finishAnimation();
       }
 
