@@ -7,54 +7,82 @@ var Box = require('../components/box');
 var EmojiSpan = require('../components/emoji-span');
 
 var CATS = ['😸', '😹', '😺', '😻', '😼', '😽', '😾', '😿', '🙀'];
-var FOODS = ['🍅', '🍆', '🍇', '🍈', '🍉', '🍊', '🍌', '🍍', '🍎', '🍏', '🍑', '🍒', '🍓', '🍔', '🍕', '🍖', '🍗'];
-
+var FOODS = [
+  '🍅',
+  '🍆',
+  '🍇',
+  '🍈',
+  '🍉',
+  '🍊',
+  '🍌',
+  '🍍',
+  '🍎',
+  '🍏',
+  '🍑',
+  '🍒',
+  '🍓',
+  '🍔',
+  '🍕',
+  '🍖',
+  '🍗',
+];
 
 var Animations = {
   // Register these with UI Pack so that we can use stagger later.
   In: velocityHelpers.registerEffect({
     calls: [
-      [{
-        transformPerspective: [ 800, 800 ],
-        transformOriginX: [ '50%', '50%' ],
-        transformOriginY: [ '100%', '100%' ],
-        marginBottom: 0,
-        opacity: 1,
-        rotateX: [0, 130],
-      }, 1, {
-        easing: 'ease-out',
-        display: 'block',
-      }]
+      [
+        {
+          transformPerspective: [800, 800],
+          transformOriginX: ['50%', '50%'],
+          transformOriginY: ['100%', '100%'],
+          marginBottom: 0,
+          opacity: 1,
+          rotateX: [0, 130],
+        },
+        1,
+        {
+          easing: 'ease-out',
+          display: 'block',
+        },
+      ],
     ],
   }),
 
   Out: velocityHelpers.registerEffect({
     calls: [
-      [{
-        transformPerspective: [ 800, 800 ],
-        transformOriginX: [ '50%', '50%' ],
-        transformOriginY: [ '0%', '0%' ],
-        marginBottom: -30,
-        opacity: 0,
-        rotateX: -70,
-      }, 1, {
-        easing: 'ease-out',
-        display: 'block',
-      }]
+      [
+        {
+          transformPerspective: [800, 800],
+          transformOriginX: ['50%', '50%'],
+          transformOriginY: ['0%', '0%'],
+          marginBottom: -30,
+          opacity: 0,
+          rotateX: -70,
+        },
+        1,
+        {
+          easing: 'ease-out',
+          display: 'block',
+        },
+      ],
     ],
   }),
 };
 
+const newItem = i => {
+  return {
+    title: [_.sample(CATS)].concat(_.sample(FOODS, _.random(1, 4))).join(' '),
+    i,
+  };
+};
+
 class ScrollingGroup extends React.Component {
   state = {
-    itemCount: 0,
-    items: [],
+    itemCount: 1,
+    items: [newItem(0)],
     duration: 500,
   };
-
-  componentWillMount() {
-    this.whenAddButtonClicked();
-  }
 
   whenAddButtonClicked = () => {
     this.addRows(1);
@@ -64,19 +92,15 @@ class ScrollingGroup extends React.Component {
     this.addRows(5);
   };
 
-  whenOptionClicked = (event) => {
+  whenOptionClicked = event => {
     this.setState({ duration: parseInt(event.target.value) });
   };
 
-  addRows = (count) => {
+  addRows = count => {
     var items = this.state.items;
 
     for (var i = 0; i < count; ++i) {
-      var item = {
-        title: [_.sample(CATS)].concat(_.sample(FOODS, _.random(1, 4))).join(' '),
-        i: this.state.itemCount + i,
-      };
-
+      var item = newItem(i + this.state.itemCount);
       items = [item].concat(items);
     }
 
@@ -87,15 +111,20 @@ class ScrollingGroup extends React.Component {
   };
 
   render() {
-    var rows = this.state.items.map(function (item, i, arr) {
+    var rows = this.state.items.map(function(item, i, arr) {
       var itemStyle = {
         width: 150,
         padding: '0 10px',
         lineHeight: '30px',
-        backgroundColor: (item.i % 2 == 0) ? Box.COLORS.backColor : Box.COLORS.underneathColor,
+        backgroundColor:
+          item.i % 2 == 0 ? Box.COLORS.backColor : Box.COLORS.underneathColor,
       };
 
-      return (<div key={item.i} style={itemStyle}><EmojiSpan>{item.title}</EmojiSpan></div>);
+      return (
+        <div key={item.i} style={itemStyle}>
+          <EmojiSpan>{item.title}</EmojiSpan>
+        </div>
+      );
     });
 
     var groupStyle = {
@@ -129,17 +158,37 @@ class ScrollingGroup extends React.Component {
           <button onClick={this.whenAdd5ButtonClicked}>Add 5 Rows</button>
         </div>
 
-        <VelocityTransitionGroup component="div" className="flex-1" style={groupStyle} enter={enterAnimation} leave={leaveAnimation}>
+        <VelocityTransitionGroup
+          component="div"
+          className="flex-1"
+          style={groupStyle}
+          enter={enterAnimation}
+          leave={leaveAnimation}
+        >
           {rows}
         </VelocityTransitionGroup>
 
-        <form style={{fontSize: 12}}>
+        <form style={{ fontSize: 12 }}>
           <label>
-            <input type="radio" name="speed" value={500} checked={this.state.duration === 500} onChange={this.whenOptionClicked}/> Fast
+            <input
+              type="radio"
+              name="speed"
+              value={500}
+              checked={this.state.duration === 500}
+              onChange={this.whenOptionClicked}
+            />{' '}
+            Fast
           </label>
           &nbsp;
           <label>
-            <input type="radio" name="speed" value={2000} checked={this.state.duration === 2000} onChange={this.whenOptionClicked}/> Slow
+            <input
+              type="radio"
+              name="speed"
+              value={2000}
+              checked={this.state.duration === 2000}
+              onChange={this.whenOptionClicked}
+            />{' '}
+            Slow
           </label>
         </form>
       </div>
