@@ -7,14 +7,7 @@ var EmojiSpan = require('../components/emoji-span');
 
 var PALS = ['👫', '👬', '👭'];
 
-var EFFECTS = [
-  'bounce',
-  'shake',
-  'flash',
-  'pulse',
-  'swing',
-  'tada',
-];
+var EFFECTS = ['bounce', 'shake', 'flash', 'pulse', 'swing', 'tada'];
 
 class TriggerBox extends React.Component {
   state = {
@@ -22,6 +15,8 @@ class TriggerBox extends React.Component {
     chain: 'stop',
     palIndex: 0,
   };
+
+  velocityRef = React.createRef();
 
   whenClicked = () => {
     this.setState({
@@ -33,17 +28,17 @@ class TriggerBox extends React.Component {
       stop: this.state.chain === 'stop',
     };
 
-    this.refs.velocity.runAnimation(opts);
+    this.velocityRef.current.runAnimation(opts);
   };
 
-  whenSelectChanged = (evt) => {
+  whenSelectChanged = evt => {
     this.setState({
       effect: evt.target.value,
       isIn: true,
     });
   };
 
-  whenOptionChanged = (evt) => {
+  whenOptionChanged = evt => {
     this.setState({ chain: evt.target.value });
   };
 
@@ -53,31 +48,62 @@ class TriggerBox extends React.Component {
     return (
       <div className="flex-box flex-column flex-1 align-items-center">
         <div>
-          <select value={this.state.effect} onChange={this.whenSelectChanged}>{this.renderEffects()}</select>
+          <select value={this.state.effect} onChange={this.whenSelectChanged}>
+            {this.renderEffects()}
+          </select>
         </div>
         <div className="flex-1">
           {/*
             Use of key here keeps the component (and its set styles) from persisting across effects.
             Avoids flashing when switching effects.
           */}
-          <VelocityComponent ref="velocity" key={this.state.effect} animation={animation}>
+          <VelocityComponent
+            ref={this.velocityRef}
+            key={this.state.effect}
+            animation={animation}
+          >
             <Box onClick={this.whenClicked} instruction="Click!">
               <EmojiSpan size={72}>{PALS[this.state.palIndex]}</EmojiSpan>
             </Box>
           </VelocityComponent>
         </div>
         <div style={{ fontSize: 12 }}>
-          <h4 style={{ marginBottom: 4, fontWeight: 'bold', textAlign: 'center'}}>Interruption Behavior</h4>
+          <h4
+            style={{ marginBottom: 4, fontWeight: 'bold', textAlign: 'center' }}
+          >
+            Interruption Behavior
+          </h4>
           <label>
-            <input type="radio" name="chain" value="stop" checked={this.state.chain === 'stop'} onChange={this.whenOptionChanged}/> Stop
+            <input
+              type="radio"
+              name="chain"
+              value="stop"
+              checked={this.state.chain === 'stop'}
+              onChange={this.whenOptionChanged}
+            />{' '}
+            Stop
           </label>
           &nbsp;&nbsp;
           <label>
-            <input type="radio" name="chain" value="finish" checked={this.state.chain === 'finish'} onChange={this.whenOptionChanged}/> Finish
+            <input
+              type="radio"
+              name="chain"
+              value="finish"
+              checked={this.state.chain === 'finish'}
+              onChange={this.whenOptionChanged}
+            />{' '}
+            Finish
           </label>
           &nbsp;&nbsp;
           <label>
-            <input type="radio" name="chain" value="queue" checked={this.state.chain === 'queue'} onChange={this.whenOptionChanged}/> Queue
+            <input
+              type="radio"
+              name="chain"
+              value="queue"
+              checked={this.state.chain === 'queue'}
+              onChange={this.whenOptionChanged}
+            />{' '}
+            Queue
           </label>
         </div>
       </div>
@@ -85,8 +111,12 @@ class TriggerBox extends React.Component {
   }
 
   renderEffects = () => {
-    return EFFECTS.map(function (effect) {
-      return (<option key={effect} value={effect}>{s.titleize(s.humanize(effect))}</option>);
+    return EFFECTS.map(function(effect) {
+      return (
+        <option key={effect} value={effect}>
+          {s.titleize(s.humanize(effect))}
+        </option>
+      );
     });
   };
 }
